requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jose==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.4.0
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.40.0
tabulate==0.9.0
tiktoken==0.8.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yoyo-migrations==9.0.0
zipp==3.20.2
